+++ manager_umbrella/apps/manager/priv/repo/migrations/20220822175246_create_exams.exs
@@ -4,7 +4,7 @@ defmodule Manager.Repo.Migrations.CreateExams do
   def change do
     create table(:exams) do
       add :name, :string
-      add :mark, :integer
+      add :mark, :float
       add :type, :string
       add :status, :string
       add :subject_id, references(:subjects, on_delete: :nothing)
